Revert "kill int32_t"
This reverts commit 260ff2dd11afc28da20b480830fa36478e4e4b08.

+++ src/cast_core/test/CrudActor_test.cpp
@@ -914,7 +914,7 @@ TEST_CASE_METHOD(MongoTestFixture,
             REQUIRE(std::string(collection) == "test");
 
             REQUIRE(countEvent["pipeline"][0]["$match"].get_document().view() ==
-                    BasicBson::make_document(BasicBson::kvp("a", int64_t(1))));
+                    BasicBson::make_document(BasicBson::kvp("a", 1)));
         } catch (const std::exception& e) {
             auto diagInfo = boost::diagnostic_information(e);
             INFO("CAUGHT " << diagInfo);
